requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
# Editable install with no version control (demented-discord-bot==2.0.0)
-e c:\servers\discord bots\demented
discord==2.3.2
discord.py==2.5.2
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nltk==3.9.1
numpy==2.3.1
packaging==25.0
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.20.1
